+++ src/app/auth/callback/page.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect } from "react";
 
-import { useAuth } from "@/hooks/auth/useAuth";
+import { useAuth } from "@/hooks/auth/useAuthApi";
 
 import LoadingLottie from "@/components/common/LoadingLottie";
