+++ src/app/viewing/reservation/[id]/confirm/page.tsx
@@ -9,7 +9,7 @@ import { useQueryClient } from "@tanstack/react-query";
 
 import { createViewing } from "@/apis/viewingApi";
 
-import { usePropertyList } from "@/hooks/property/useProperty";
+import { usePropertyList } from "@/hooks/property/usePropertyApi";
 
 import { formatDateTime, getTimeLabel } from "@/utils/formatter/dateFormatter";
