+++ src/components/viewings/ViewingConfirmedSection.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 
 import { groupBy } from "lodash";
 
-import { useCancelViewing } from "@/hooks/viewing/useViewing";
+import { useCancelViewing } from "@/hooks/viewing/useViewingApi";
 
 import { calculateDday } from "@/utils/formatter/dateFormatter";
