+++ src/components/viewings/ViewingNoteSection.tsx
@@ -3,7 +3,7 @@ import { forwardRef, useEffect, useImperativeHandle, useState } from "react";
 import { getPropertyNotePresignedUrl } from "@/apis/s3UploadApi";
 import { putViewingPropertyNotes } from "@/apis/viewingApi";
 
-import { useViewingDetail } from "@/hooks/viewing/useViewing";
+import { useViewingDetail } from "@/hooks/viewing/useViewingApi";
 
 import { uploadFilesToPresignedUrls } from "@/utils/images/uploadFilesToPresignedUrls";
